+++ R/tt.R
@@ -40,7 +40,9 @@
 #' 
 #' tt(x, notes = "Hello World!")
 #'
-#' tt(x, notes = list("*" = list(i = 0:1, j = 2, text = "Hello World!")))
+#' fn <- list(i = 0:1, j = 2, text = "Hello World!")
+#' tab <- tt(x, notes = list("*" = fn))
+#' print(tab, "latex")
 #' 
 #' @export
 tt <- function(x,